Feat: circuit breaker 추가
- circuit breaker 추가
- controller 리팩토링 serviceImpl 수정

+++ journey-service/src/main/java/com/ssafy/journeymate/journeyservice/client/CategoryServiceClient.java
@@ -11,9 +11,9 @@ import org.springframework.web.bind.annotation.PathVariable;
 public interface CategoryServiceClient {
 
     @GetMapping("/category-service/{categoryId}")
-    CategoryGetRes getCategory(@PathVariable int categoryId);
+    CategoryGetRes getCategory(@PathVariable Long categoryId);
 
     @GetMapping("/category-service/{categoryId}/items")
-    List<ItemGetRes> getCategoryItems(@PathVariable int categoryId);
+    List<ItemGetRes> getCategoryItems(@PathVariable Long categoryId);
 
 }
